Split of MQTT and MQTT Auto Discovery code into new hardware type
