+++ CMakeLists.txt
@@ -254,6 +254,7 @@ hardware/Meteorologisk.cpp
 hardware/Meteostick.cpp
 hardware/MochadTCP.cpp
 hardware/MQTT.cpp
+hardware/MQTTAutoDiscover.cpp
 hardware/MultiFun.cpp
 hardware/MySensorsBase.cpp
 hardware/MySensorsSerial.cpp
