+++ main/RFXNames.h
@@ -221,6 +221,7 @@ enum _eHardwareTypes {
 	HTYPE_Mercedes,				//122
 	HTYPE_AirconWithMe,         //123
 	HTYPE_TeleinfoMeterTCP,		//124
+	HTYPE_MQTTAutoDiscovery,	//125
 	HTYPE_END
 };
