+++ msbuild/domoticz.vcxproj
@@ -611,6 +611,7 @@
     <ClInclude Include="..\hardware\HardwareCereal.h" />
     <ClInclude Include="..\hardware\Honeywell.h" />
     <ClInclude Include="..\hardware\Meteorologisk.h" />
+    <ClInclude Include="..\hardware\MQTTAutoDiscover.h" />
     <ClInclude Include="..\hardware\NestOAuthAPI.h" />
     <ClInclude Include="..\hardware\OctoPrintMQTT.h" />
     <ClInclude Include="..\hardware\plugins\PythonObjectEx.h" />
@@ -915,6 +916,7 @@
     <ClCompile Include="..\hardware\GpioPin.cpp" />
     <ClCompile Include="..\hardware\Honeywell.cpp" />
     <ClCompile Include="..\hardware\Meteorologisk.cpp" />
+    <ClCompile Include="..\hardware\MQTTAutoDiscover.cpp" />
     <ClCompile Include="..\hardware\NestOAuthAPI.cpp" />
     <ClCompile Include="..\hardware\OctoPrintMQTT.cpp" />
     <ClCompile Include="..\hardware\plugins\PythonObjectEx.cpp" />
